moved Openstack Infra views to openstack_infra.py
